The initial guess system is completely reorganized. It is now possible to guess the gain for each type of channel (signal, pump, ase, raman) as well as the rough shape of the function (linear or exponential).
The result of a previous simulation can be used as the initial guess as
well to perform fast parameter sweeps. The initial guess array can also
be constructed totally outside pyfiberamp, though this is not yet
documented.

One example and one test case have been added for the parameter sweep
functionality.